Nicholas Loehrke Final Code/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asttokens==2.4.1
black==24.10.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
emoji==2.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
executing==2.1.0
fasttext-wheel==0.9.2
filelock==3.13.1
fonttools==4.54.1
fsspec==2024.2.0
grpcio==1.66.2
idna==3.10
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.28.2
pure_eval==0.2.3
pybind11==2.13.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tensorboard==2.18.0
tensorboard-data-server==0.7.2
thinc==8.3.2
threadpoolctl==3.5.0
tomli==2.0.2
torch==2.4.1+cu124
torchaudio==2.4.1+cu124
torchvision==0.19.1+cu124
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.0.4
wrapt==1.16.0
